FullyFunctionalHomeAndTourEnvironmentTeleportation
A user now teleports seamlessly between HomeEnvironment and TourEnvironment. This involves clicking the home button in order to teleport from home to tour or tour to home as well as grabbing a building to teleport to tour. The home button from home to tour does not work if a user has not visited any location yet.